src/main/java/com/anil/securenotes/controllers/WelcomeController.java
package com.anil.securenotes.controllers;


import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api")
public class WelcomeController {

    @GetMapping(value = "/admin")
    public String admin() {
        return "admin";
    }


    @GetMapping(value = "/user", produces = MediaType.APPLICATION_JSON_VALUE)
    public String user() {
        return "user";
    }
}
